chore(tests) refactor checkEthTx
